+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 * Fixed Integer/Decimal field serializations [#125](https://github.com/contentful/contentful.rb/issues/125)
+* Fixed File coercion for Localized Assets [#129](https://github.com/contentful/contentful.rb/issues/129)
 
 ## 2.0.0